requirements.txt
discord.py
riotwatcher
